Updated rna_synthesizer. Removed rna_synthesizer_n.
